remove freezed & sanitize disciplines/details
