--- lib/src/application/discipline/details/event.dart
@@ -1,6 +1,0 @@
-part of 'bloc.dart';
-
-@freezed
-class DisciplineDetailsEvent with _$DisciplineDetailsEvent {
-  const factory DisciplineDetailsEvent.started() = _Started;
-}
