--- lib/src/application/discipline/details/state.dart
@@ -1,6 +1,0 @@
-part of 'bloc.dart';
-
-@freezed
-class DisciplineDetailsState with _$DisciplineDetailsState {
-  const factory DisciplineDetailsState() = _DisciplineDetailsState;
-}
