+++ lib/src/presentation/disciplines/details/discipline_details_page.dart
@@ -1,15 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../application/discipline/details/bloc.dart';
 import '../../../domain/discipline.dart';
-import 'widgets/discipline_details_view.dart';
+import '../../app/router.dart';
+import '../../pages.dart';
 
 class DisciplineDetailsPage extends StatelessWidget {
-  const DisciplineDetailsPage({
-    super.key,
-    required this.discipline,
-  });
+  const DisciplineDetailsPage({super.key, required this.discipline});
 
   final Discipline discipline;
 
@@ -21,12 +17,91 @@ class DisciplineDetailsPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return BlocProvider<DisciplineDetailsBloc>(
-      create: (_) {
-        final bloc = DisciplineDetailsBloc(discipline: discipline);
-        return bloc..add(const DisciplineDetailsEvent.started());
-      },
-      child: DisciplineDetailsView(discipline: discipline),
+    return DisciplineDetailsView(discipline: discipline);
+  }
+}
+
+class DisciplineDetailsView extends StatelessWidget {
+  const DisciplineDetailsView({super.key, required this.discipline});
+
+  final Discipline discipline;
+
+  static const List<Widget> tabs = [
+    Tab(text: 'Alunos'),
+    Tab(text: 'Chamadas'),
+  ];
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+
+    return DefaultTabController(
+      length: tabs.length,
+      child: Scaffold(
+        appBar: AppBar(
+          titleSpacing: 0,
+          title: Text(discipline.name),
+          titleTextStyle: theme.textTheme.titleMedium,
+          bottom: const TabBar(tabs: tabs),
+        ),
+        body: TabBarView(
+          children: [
+            StudentsOverviewPage(discipline: discipline),
+            AttendancesOverviewPage(discipline: discipline),
+          ],
+        ),
+        floatingActionButtonLocation: FloatingActionButtonLocation.endContained,
+        floatingActionButton: FloatingActionButton(
+          heroTag: '_', // Temporary fix for scaffold paint issue
+          child: const Icon(Icons.add_task_rounded),
+          onPressed: () => AppRouter.showAttendanceForm(context, discipline),
+        ),
+        bottomNavigationBar: DisciplineDetailsBottomBar(discipline: discipline),
+      ),
+    );
+  }
+}
+
+class DisciplineDetailsBottomBar extends StatelessWidget {
+  const DisciplineDetailsBottomBar({super.key, required this.discipline});
+
+  final Discipline discipline;
+
+  @override
+  Widget build(BuildContext context) {
+    return BottomAppBar(
+      padding: const EdgeInsets.symmetric(vertical: 12.0, horizontal: 8.0),
+      child: Row(
+        children: [
+          IconButton(
+            icon: const Icon(Icons.format_list_bulleted_add),
+            tooltip: 'Importar Alunos',
+            onPressed: () async {
+              await showModalBottomSheet<void>(
+                context: context,
+                isScrollControlled: true,
+                builder: (_) => StudentsImportPage(discipline: discipline),
+              );
+            },
+          ),
+          IconButton(
+            icon: const Icon(Icons.manage_accounts),
+            tooltip: 'Editar Alunos',
+            onPressed: () => AppRouter.showStudentsForm(context, discipline),
+          ),
+          IconButton(
+            icon: const Icon(Icons.publish_rounded),
+            tooltip: 'Exportar Chamadas',
+            onPressed: () async {
+              await showModalBottomSheet<void>(
+                context: context,
+                isScrollControlled: true,
+                builder: (_) => DisciplineExportPage(discipline: discipline),
+              );
+            },
+          ),
+        ],
+      ),
     );
   }
 }
