--- lib/src/presentation/disciplines/details/widgets/body/_attendances_tab.dart
@@ -1,11 +1,0 @@
-part of 'body.dart';
-
-class AttendancesTab extends StatelessWidget {
-  const AttendancesTab({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final bloc = context.watch<DisciplineDetailsBloc>();
-    return AttendancesOverviewPage(discipline: bloc.discipline);
-  }
-}
